Refactor: Centralize voices label and shared-team filtering in VoicesViewModel
- Move `filterNewsByLabel` and `extractSharedTeamName` logic from `VoicesFragment` to `VoicesViewModel`.
- Move `JsonUtils` and `Constants.LABELS` parsing out of the UI layer to the ViewModel.
- Keep spinner wiring and adapter `submitList` calls in the fragment to avoid broad UI rewrites.
- Ensure the selected label filtering handles "All", "Shared Chat", pre-defined labels, and dynamic shared team names properly.
