+++ app/src/main/java/org/ole/planet/myplanet/ui/voices/VoicesViewModel.kt
@@ -10,8 +10,13 @@ import org.ole.planet.myplanet.model.RealmMyLibrary
 import org.ole.planet.myplanet.model.RealmUser
 import org.ole.planet.myplanet.repository.TeamsRepository
 import org.ole.planet.myplanet.repository.UserRepository
+import com.google.gson.JsonArray
+import org.ole.planet.myplanet.model.RealmNews
 import org.ole.planet.myplanet.repository.VoicesRepository
+import org.ole.planet.myplanet.services.VoicesLabelManager
+import org.ole.planet.myplanet.utils.Constants
 import org.ole.planet.myplanet.utils.DispatcherProvider
+import org.ole.planet.myplanet.utils.JsonUtils
 
 @HiltViewModel
 class VoicesViewModel @Inject constructor(
@@ -80,4 +85,73 @@ class VoicesViewModel @Inject constructor(
         }
     }
 
+    fun filterNewsByLabel(
+        newsList: List<RealmNews?>,
+        selectedLabel: String
+    ): Pair<List<String>, List<RealmNews?>> {
+        val labelDisplayToValue = mutableMapOf<String, String>()
+        val allLabels = mutableSetOf<String>()
+        allLabels.add("All")
+
+        Constants.LABELS.forEach { (labelName, labelValue) ->
+            allLabels.add(labelName)
+            labelDisplayToValue[labelName] = labelValue
+        }
+
+        allLabels.add("Shared Chat")
+
+        newsList.forEach { news ->
+            val sharedTeamName = extractSharedTeamName(news)
+            if (sharedTeamName.isNotEmpty()) {
+                allLabels.add(sharedTeamName)
+            }
+
+            news?.labels?.forEach { label ->
+                val labelName = Constants.LABELS.entries.find { it.value == label }?.key
+                    ?: VoicesLabelManager.formatLabelValue(label)
+                allLabels.add(labelName)
+                labelDisplayToValue.putIfAbsent(labelName, label)
+            }
+        }
+
+        val displayLabels = allLabels.sorted()
+
+        val filteredList = if (selectedLabel == "All") {
+            newsList
+        } else {
+            newsList.filter { news ->
+                when {
+                    selectedLabel == "Shared Chat" -> {
+                        news?.chat == true || news?.viewableBy.equals("community", ignoreCase = true)
+                    }
+                    labelDisplayToValue.containsKey(selectedLabel) -> {
+                        val labelValue = labelDisplayToValue[selectedLabel]
+                        news?.labels?.contains(labelValue) == true
+                    }
+                    else -> {
+                        extractSharedTeamName(news) == selectedLabel
+                    }
+                }
+            }
+        }
+
+        return Pair(displayLabels, filteredList)
+    }
+
+    private fun extractSharedTeamName(news: RealmNews?): String {
+        if (!news?.viewIn.isNullOrEmpty()) {
+            try {
+                val ar = JsonUtils.gson.fromJson(news.viewIn, JsonArray::class.java)
+                if (ar.size() > 1) {
+                    val ob = ar[0].asJsonObject
+                    if (ob.has("name") && !ob.get("name").isJsonNull) {
+                        return ob.get("name").asString
+                    }
+                }
+            } catch (e: Exception) {
+                e.printStackTrace()
+            }
+        }
+        return ""
+    }
 }
